flatten Penalty.vue to 5 vue files
