+++ package.json
@@ -1,7 +1,7 @@
 {
     "name": "mercuryland",
     "private": true,
-    "version": "2.4.3_20240923",
+    "version": "2.5.0_20240927",
     "type": "module",
     "scripts": {
         "dev": "vite",
