+++ src/components/penalty/Syntax.vue
